Split options and use shared classes for shape cells
