+++ Samples/PagingLayoutSamples/Modules/LayoutDesigner/Options/cell/LayoutDesignerOptionCellViewModel.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import UIKit
 
 class LayoutDesignerOptionCellViewModel {
