+++ Samples/PagingLayoutSamples/Modules/Shapes/ShapesViewModel.swift
@@ -7,14 +7,33 @@
 //
 
 import UIKit
+import CollectionViewPagingLayout
 
 class ShapesViewModel {
     
+    // MARK: Static Properties
+    
+    private(set) static var scaleOptions: ScaleTransformViewOptions = .init()
+    private(set) static var stackOptions: StackTransformViewOptions = .init()
+    private(set) static var snapshotOptions: SnapshotTransformViewOptions = .init()
+    
+    
     // MARK: Properties
     
-    var selectedLayout: LayoutTypeCellViewModel
+    var selectedLayout: LayoutTypeCellViewModel {
+        didSet {
+            if let options = selectedLayout.layout.scaleOptions {
+                ShapesViewModel.scaleOptions = options
+            } else if let options = selectedLayout.layout.stackOptions {
+                ShapesViewModel.stackOptions = options
+            } else if let options = selectedLayout.layout.snapshotOptions {
+                ShapesViewModel.snapshotOptions = options
+            }
+        }
+    }
     let layoutTypeViewModels: [LayoutTypeCellViewModel]
     let showBackButton: Bool
+    
     
     // MARK: Lifecycle
     
@@ -34,136 +53,114 @@ extension ShapesViewModel {
               iconName: "scale_invertedcylinder",
               title: "Scale",
               subtitle: "Inverted Cylinder",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)]),
-              cellClass: ScaleInvertedCylinderShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)])),
         .init(layout: .scaleCylinder,
               iconName: "scale_cylinder",
               title: "Scale",
               subtitle: "Cylinder",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)]),
-              cellClass: ScaleCylinderShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)])),
         .init(layout: .scaleCoverFlow,
               iconName: "scale_coverflow",
               title: "Scale",
               subtitle: "Cover Flow",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)]),
-              cellClass: ScaleCoverFlowShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)])),
         .init(layout: .scaleLinear,
               iconName: "scale_normal",
               title: "Scale",
               subtitle: "Linear",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)]),
-              cellClass: ScaleLinearShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)])),
         .init(layout: .scaleEaseIn,
               iconName: "scale_normal",
               title: "Scale",
               subtitle: "EaseIn",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)]),
-              cellClass: ScaleEaseInShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)])),
         .init(layout: .scaleEaseOut,
               iconName: "scale_normal",
               title: "Scale",
               subtitle: "EaseOut",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)]),
-              cellClass: ScaleEaseOutShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)])),
         .init(layout: .scaleRotary,
               iconName: "scale_rotary",
               title: "Scale",
               subtitle: "Rotary",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)]),
-              cellClass: ScaleRotaryShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)])),
         .init(layout: .scaleBlur,
               iconName: "scale_blur",
               title: "Scale",
               subtitle: "Blurry",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)]),
-              cellClass: ScaleBlurShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 1, green: 0.4274509804, blue: 0.4, alpha: 1), #colorLiteral(red: 1, green: 0.7803921569, blue: 0, alpha: 1)])),
         
         .init(layout: .stackTransparent,
               iconName: "stack_transparent",
               title: "Stack",
               subtitle: "Transparent",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)]),
-              cellClass: TransparentStackShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)])),
         .init(layout: .stackPerspective,
               iconName: "stack_prespective",
               title: "Stack",
               subtitle: "Perspective",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)]),
-              cellClass: PerspectiveStackShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)])),
         .init(layout: .stackRotary,
               iconName: "stack_rotary",
               title: "Stack",
               subtitle: "Rotary",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)]),
-              cellClass: RotaryStackShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)])),
         .init(layout: .stackBlur,
               iconName: "stack_blur",
               title: "Stack",
               subtitle: "Blur",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)]),
-              cellClass: BlurStackShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)])),
         .init(layout: .stackVortex,
               iconName: "stack_vortex",
               title: "Stack",
               subtitle: "Vortex",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)]),
-              cellClass: VortexStackShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)])),
         .init(layout: .stackReverse,
               iconName: "stack_reverse",
               title: "Stack",
               subtitle: "Reverse",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)]),
-              cellClass: ReverseStackShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.3058823529, green: 1, blue: 0.9490196078, alpha: 1), #colorLiteral(red: 0.1137254902, green: 0.4156862745, blue: 0.6666666667, alpha: 1)])),
         
         .init(layout: .snapshotGrid,
               iconName: "snapshot_grid",
               title: "Snapshot",
               subtitle: "Grid",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]),
-              cellClass: GridSnapshotShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)])),
         .init(layout: .snapshotSpace,
               iconName: "snapshot_space",
               title: "Snapshot",
               subtitle: "Space",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]),
-              cellClass: SpaceSnapshotShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)])),
         .init(layout: .snapshotChess,
               iconName: "snapshot_chess",
               title: "Snapshot",
               subtitle: "Chess",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]),
-              cellClass: ChessSnapshotShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)])),
         .init(layout: .snapshotTiles,
               iconName: "snapshot_tiles",
               title: "Snapshot",
               subtitle: "Tiles",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]),
-              cellClass: TilesSnapshotShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)])),
         .init(layout: .snapshotLines,
               iconName: "snapshot_lines",
               title: "Snapshot",
               subtitle: "Lines",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]),
-              cellClass: LinesSnapshotShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)])),
         .init(layout: .snapshotBars,
               iconName: "snapshot_bars",
               title: "Snapshot",
               subtitle: "Bars",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]),
-              cellClass: BarsSnapshotShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)])),
         .init(layout: .snapshotPuzzle,
               iconName: "snapshot_puzzle",
               title: "Snapshot",
               subtitle: "Puzzle",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]),
-              cellClass: PuzzleSnapshotShapeCollectionViewCell.self),
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)])),
         .init(layout: .snapshotFade,
               iconName: "snapshot_fade",
               title: "Snapshot",
               subtitle: "Fade",
-              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]),
-              cellClass: FadeSnapshotShapeCollectionViewCell.self)
+              cardViewModels: generateCardViewModels(colors: [#colorLiteral(red: 0.9176470588, green: 1, blue: 0, alpha: 1), #colorLiteral(red: 0.3137254902, green: 0.8, blue: 1, alpha: 1)]))
     ]
     
     private static func generateCardViewModels(colors: [UIColor]) -> [ShapeCardViewModel] {
